Parser class name is not a good name. Renamed it to RegExp and moved parsing to a separate parser.rs file.
